Push 15
This was a major push although not the final push. It includes all the major elements of our final app and just needs a bit of polishing.
This push contains the home page with the ability to quick launch onto pages previously saved.
The push contains the general note page with multiple pages. The picker code was used to implemtent this.
We aren't aware of any bugs making this push.